Optimize PBXProj.getFileElement()
This file uses a pattern which collects dictionaries into array, which is then
searched in order for a specific key. This search pattern is similar to `??`
but works across an array instead of two parameters.

This change to `getFileElement()` replaces this pattern with `??` chaining
because profiling showed this pattern is poorly optimized by swift, and thus
`getFileElement()` is a large hot spot.

In particular, for project generation with XcodeGen, the project I work went
from 9-10s on my machine, to 1.5-2s. Total memory allocations went down from
2.6gb to 0.5gb.

In Instruments, the category
`_HashableTypedNativeDictionaryStorage<String, PBXFileElement>` had an
allocation count of 61595 for a sum total of 2.13gb.  With this change,
Instruments shows no such allocations.

All these allocations were wasting cpu as well. With this change,
`_swift_release_` and `_swift_retain_` were reduced to 85ms each, down from
2.3s and 1.9s respectively.