Refactor the XCSchemeIntegrationSpec Test Case
In preparation for adding additional scheme integration tests, remove the subject property and change test_init_initializesTheSchemeCorrectly() to load/assert the scheme directly (the property was otherwise unused). This gives us one independent function for verifying reading and one independent function for verifying writing.

Originally I called the assert function in the initModel closure in the test_write() method but decided against that because I didn't like the read test being a side effect of the write test.