Relax Scheme Parsing
Xcode accepts just about any scheme file, filling the model with default values if necessary. Update the XCScheme XML initialization to better match Xcode's behavior.

MinimalInformation.xcscheme was taken from a large codebase with 279 Xcode projects. After simply opening the scheme in the Scheme Editor and closing the sheet, Xcode rewrites the file with default values. These default values were integrated in to the XCScheme loading code as fallbacks for when the attribute or element doesn't exist.

Also note that the "buildableProductRunnable" property/element isn't required. It's possible to select "None" for the executable in the Scheme Editor, in which case Xcode won't generate the element.

Also update the XCScheme integration test to test reading and writing the underspecified scheme to ensure it can be successfully loaded with the correct default values.