Optimization: Only escape CommentedString when necessary
Every call to `replacingOccurrences()` creates a new copy of the string, even
when no replacements occur. For `CommentedString.validString`, this means 4
copies of the string are made every time, even in a common case where no
escaping is required. This code is called frequently and is a hotspot.

Additionally, Instruments is showing an allocation for each swift `String`
bridged to objective-c. This means each avoided `replacingOccurrences()` call is
also saving 2 string allocations for the pattern and replacement strings. The
briding cost is why the `contains()` checks are explicitly typed to `Character`.
Without the explicit type, the `String` version of `contains()` is called.