Add Support for Build Carbon Resources Build Phase
Some macOS apps still build legacy Carbon resources. Fortunately the Xcode project file models this in the same way as the abstract PBXBuildPhase so implementation of the build phase is as easy as adding a subclass, updating the BuildPhase enum, updating PBXProj, and adding decoding/encoding support.

In addition to updating the tests, verified xcproj correctly round trips a project with this build phase.