Allow custom Source Tree Specifications
An esoteric feature of Xcode allows people to specify a custom variable for the source tree (e.g. DERIVED_FILE_DIR). Convert PBXSourceTree to an enum with an associated value so we can load this information and update the decoding site to support any value.

With this change, loading a PBXSourceTree from a string always succeeds, which identified a number of obsolete tests that were attempting to check for a throw when a value was missing from the object's encoded representation. I removed those tests.