Load Xcode Project-Relative XCConfig Files
If an xcconfig file includes another xcconfig file with a relative path, Xcode will first attempt to load the included file relative to the including xcconfig's path. If that fails, Xcode will attempt to load the included xcconfig file relative to the Xcode project. Add support for the latter form of include resolution.