Add productInstallPath Property to PBXNativeTarget
The productInstallPath is a PBXNativeTarget-specific property. Add support for reading and writing the property so it’s not lost when round-tripping projects that use it with xcproj.

It’s not clear to me if this is still accessible in the Xcode UI or if it’s only a carryover from legacy projects.