Fix Writing of showEnvVarsInLog
Xcode writes showEnvVarsInLog to the pbxproj file if it’s set to false. If it’s true it’s omitted. Update xcproj to use the same behavior to eliminate diffs when round tripping Xcode projects.

This requires defaulting to true when decoding the project file if the key is omitted. I also updated the default parameter in the initializer to true for good measure.