Add User Comment Support
Older versions of Xcode (Xcode 3 and prior) supported user comments on build phases and targets. Add support for loading and writing comments from old project files so the project is left unchanged after a round trip.

The class name PBXContainerItem was chosen to match the class name used by Xcode which introduces the comments property in to the hierarchy.