Support Multiple Project Roots in PBXProject
Some Xcode projects have multiple project roots. In this case, Xcode writes an array of strings to the projectRoots key instead of a single string to the projectRoot key.

Follow Xcode’s lead in providing a single projectRoots property. Deprecate projectRoot and re-implement it around the projectRoots array.

Unfortunately this introduces a little unusual complexity in decoding and plist generation but it maintains consistent behavior with Xcode.