Add file reference identifier
The file reference identifier may have been present in earlier versions of xcproj but was no longer exposed. This value is valuable for matching up files to specific references in other places in the project.

e.g. getting the variantGroup that includes localizable `Localizable.strings` and then iterating the `fileReferences` to match up the file that matches the `Localizable.strings` variant by reference identifier.