Remove open collective content and say thanks badge
I don't think is fair pointing the say thanks badge to me personally. Any contributor in Tuist's projects deserves the recognition for their commitment for the project. For that reason I've removed the badge from the repository.
Moreover, I don't think it's a good idea to have an open collective account associated to the organization. I started building Tuist to have fun and learn a lot along the ride. It'd be a different story if I wanted to work on this full time or had to pay for expensive infrastructure. Since that's not the case, I'm removing the open collective account and references from the README.