Fix linux build and autogenerate XCScheme equality
Swift 5.4 changed something that caused the linux build to (correctly)
fail to compile an @objc declaration. We can do the same thing we did
to the PBXObject hierarchy in #585, and reimplement the dynamic equality
checking using memberwise equality functions generated by sourcery, and
a dynamic method on each XCScheme object.

Tested the generated interface with these changes:

    swift build -Xswiftc -emit-module-interface -Xswiftc -swift-version -Xswiftc 5

The only change to the interface is that RemoteRunnable now exposes an
`==` method:

```diff
+    public static func == (lhs: XcodeProj.XCScheme.RemoteRunnable, rhs: XcodeProj.XCScheme.RemoteRunnable) -> Swift.Bool
```